revealer: split some core parts out into separate file
for easier testing
