+++ electrum/plugins/revealer/qt.py
@@ -13,47 +13,26 @@ import os
 import random
 import qrcode
 import traceback
-from hashlib import sha256
 from decimal import Decimal
-from typing import NamedTuple, Optional, Dict, Tuple
 
 from PyQt5.QtPrintSupport import QPrinter
 
-from electrum.plugin import BasePlugin, hook
+from electrum.plugin import hook
 from electrum.i18n import _
-from electrum.util import to_bytes, make_dir, InvalidPassword, UserCancelled, bh2u, bfh
+from electrum.util import make_dir, InvalidPassword, UserCancelled, bh2u, bfh
 from electrum.gui.qt.util import *
 from electrum.gui.qt.qrtextedit import ScanQRTextEdit
 from electrum.gui.qt.main_window import StatusBarButton
 
-from .hmac_drbg import DRBG
+from .revealer import RevealerPlugin, VersionedSeed
 
 
-class VersionedSeed(NamedTuple):
-    version: str
-    seed: str
-    checksum: str
-
-    def get_ui_string_version_plus_seed(self):
-        version, seed = self.version, self.seed
-        assert isinstance(version, str) and len(version) == 1, version
-        assert isinstance(seed, str) and len(seed) >= 32
-        ret = version + seed
-        ret = ret.upper()
-        return ' '.join(ret[i : i+4] for i in range(0, len(ret), 4))
-
-
-class Plugin(BasePlugin):
-
-    LATEST_VERSION = '1'
-    KNOWN_VERSIONS = ('0', '1')
-    assert LATEST_VERSION in KNOWN_VERSIONS
+class Plugin(RevealerPlugin):
 
     MAX_PLAINTEXT_LEN = 189  # chars
-    SIZE = (159, 97)
 
     def __init__(self, parent, config, name):
-        BasePlugin.__init__(self, parent, config, name)
+        RevealerPlugin.__init__(self, parent, config, name)
         self.base_dir = os.path.join(config.electrum_path(), 'revealer')
 
         if self.config.get('calibration_h') is None:
@@ -168,30 +147,6 @@ class Plugin(BasePlugin):
             self.versioned_seed = versioned_seed
         self.user_input = bool(versioned_seed)
         self.next_button.setEnabled(bool(versioned_seed))
-
-    @classmethod
-    def code_hashid(cls, txt: str) -> str:
-        x = to_bytes(txt, 'utf8')
-        hash = sha256(x).hexdigest()
-        return hash[-3:].upper()
-
-    @classmethod
-    def get_versioned_seed_from_user_input(cls, txt: str) -> Optional[VersionedSeed]:
-        if len(txt) < 34:
-            return None
-        try:
-            int(txt, 16)
-        except:
-            return None
-        version = txt[0]
-        if version not in cls.KNOWN_VERSIONS:
-            return None
-        checksum = cls.code_hashid(txt[:-3])
-        if txt[-3:].upper() != checksum.upper():
-            return None
-        return VersionedSeed(version=version.upper(),
-                             seed=txt[1:-3].upper(),
-                             checksum=checksum.upper())
 
     def make_digital(self, dialog):
         self.make_rawnoise(True)
@@ -376,33 +331,6 @@ class Plugin(BasePlugin):
         self.rawnoise = rawnoise
         if create_revealer:
             self.make_revealer()
-
-    @classmethod
-    def get_noise_map(self, versioned_seed: VersionedSeed) -> Dict[Tuple[int, int], int]:
-        """Returns a map from (x,y) coordinate to pixel value 0/1, to be used as noise."""
-        w, h = self.SIZE
-        version  = versioned_seed.version
-        hex_seed = versioned_seed.seed
-        checksum = versioned_seed.checksum
-        noise_map = {}
-        if version == '0':
-            random.seed(int(hex_seed, 16))
-            for x in range(w):
-                for y in range(h):
-                    noise_map[(x, y)] = random.randint(0, 1)
-        elif version == '1':
-            prng_seed = bfh(hex_seed + version + checksum)
-            drbg = DRBG(prng_seed)
-            num_noise_bytes = 1929  # ~ w*h
-            noise_array = bin(int.from_bytes(drbg.generate(num_noise_bytes), 'big'))[2:]
-            i = 0
-            for x in range(w):
-                for y in range(h):
-                    noise_map[(x, y)] = int(noise_array[i])
-                    i += 1
-        else:
-            raise Exception(f"unexpected revealer version: {version}")
-        return noise_map
 
     def make_calnoise(self):
         random.seed(self.calibration_noise)
